# deploy/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ProfilePatchRequest": {
        "description": "PATCH /profile \u2014 correct an onboarding value. Only changed field(s) + client_request_id.",
        "properties": {
          "age": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "bodyweight_kg": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodyweight Kg"
          },
          "client_request_id": {
            "minLength": 1,
            "title": "Client Request Id",
            "type": "string"
          },
          "experience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience"
          },
          "sex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sex"
          }
        },
        "required": [
          "client_request_id"
        ],
        "title": "ProfilePatchRequest",
        "type": "object"
      },
...
      "SessionStartRequest": {
        "description": "POST /sessions \u2014 no athlete-supplied composition parameters (composition is server-owned\nand load-free; the client cannot influence it). Idempotent on client_request_id.",
        "properties": {
          "client_request_id": {
            "minLength": 1,
            "title": "Client Request Id",
            "type": "string"
          }
        },
        "required": [
          "client_request_id"
        ],
        "title": "SessionStartRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Hush v1 \u2014 advisory training intelligence. Hush recommends; the athlete decides.",
    "title": "Hush v1 API",
    "version": "v1",
    "x-hush-schema-version": 11
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/profile": {
      "get": {
        "operationId": "profile_get_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Profile Get",
        "tags": [
          "read"
        ]
      },
      "patch": {
        "operationId": "patch_profile_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Profile",
        "tags": [
          "profile"
        ]
      }
    },
    "/recommendations/{recommendation_id}/why": {
      "get": {
        "operationId": "recommendation_why_recommendations__recommendation_id__why_get",
        "parameters": [
          {
            "in": "path",
            "name": "recommendation_id",
            "required": true,
            "schema": {
              "title": "Recommendation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recommendation Why",
        "tags": [
          "read"
        ]
      }
    },
    "/sessions": {
      "get": {
        "operationId": "sessions_list_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sessions List",
        "tags": [
          "read"
        ]
      },
      "post": {
        "operationId": "start_session_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Session",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/sessions/{session_id}": {
      "get": {
        "operationId": "session_detail_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session Detail",
        "tags": [
          "read"
        ]
      }
    },
...
  }
}